fix mem size_t

+++ src/syncedmem.cpp
@@ -172,7 +172,7 @@ inline bool ShouldBorrowMem(int has, int wants) {
   return has / 2 <= wants;
 }
 
-MemBlock MemoryPool::RequestCPU(int size) {
+MemBlock MemoryPool::RequestCPU(size_t size) {
   MemBlock block;
   if (size <= kElementSize) {  // small object <= 128 bytes
     block.device = -1;
@@ -231,7 +231,7 @@ void MemoryPool::ReturnCPU(MemBlock block) {
   }
 }
 
-MemBlock MemoryPool::RequestGPU(int size, int device) {
+MemBlock MemoryPool::RequestGPU(size_t size, int device) {
   MemBlock block;
 #ifdef USE_CUDA
   GpuKey key{device, size};
